dont allow users to purchase premium if they already have it
