+++ lib/philomena_web/controllers/premium_subscription/subscribe_controller.ex
@@ -8,9 +8,16 @@ defmodule PhilomenaWeb.PremiumSubscription.SubscribeController do
   require Logger
 
   def create(conn, %{"premium_subscription_id" => plan_id}) do
-    conn
-    |> create_new_subscription(plan_id)
-    |> (&(redirect(conn, external: &1))).()
+    case User.premium?(conn.assigns.current_user) do
+      true ->
+        conn
+        |> put_flash(:info, "You already have a premium subscription")
+        |> redirect(external: Routes.activity_path(conn, :index))
+      false ->
+        conn
+        |> create_new_subscription(plan_id)
+        |> (&(redirect(conn, external: &1))).()
+    end
   end
 
   def index(conn, %{"premium_subscription_id" => id}) do
